Minor touchups, Separate Color Sequences
Refactoring across the entire project including the main plugin file and moving the RTC to a Components folder. Replaced current sequence configs with a dictionary to accept a group name as a key and have a list of strings as the sequence for the value. Now checks to see if a RT should be granted OnChangingGroup.